Block implementations and other stuff
Motivational Block: when you are within range of this block you will be
able to teleport further than your max teleport distance

Border Crossing Block: This makes it so that you cannot teleport into
the aoe of a border crossing block from outside it's aoe and vice versa

Trigger Block: When you are on this block and you select it again it
will call the activate function of the block. If you need to create new
trigger functionality all you have to do is create a new script and add
an Activate() Function to the script.

Added a triggerscript for moving blocks that turns movement off/on

Created a VR pause menu and a player prefab that includes it (For the
mobile build this pause menu must be used because the current pc pause
menu will not work.

Orginizational stuff: moved scripts into the scripts folder and
organized the different player prefabs.